requirements.txt
# Essential Requirements for LangChain Streamlit Application

# Core Web Framework
streamlit>=1.41.1
fastapi>=0.104.1

# LangChain Framework
langchain>=0.3.26
langchain-core>=0.3.74
langchain-community>=0.3.26
langchain-experimental>=0.3.4
langchain-text-splitters>=0.3.8

# LangChain LLM Integrations
langchain-openai>=0.2.14
langchain-groq>=0.2.2
langchain-google-genai>=2.1.4
# langchain-ollama>=0.3.6
langchain-huggingface>=0.1.2

# LangChain Tools & Extensions
langgraph>=0.5.3
langsmith>=0.4.5
langserve>=0.3.1

# AI/ML Core Libraries
openai>=1.75.0
groq>=0.13.1
google-generativeai>=0.8.3
anthropic>=0.40.0
cohere>=5.13.11

# Document Processing
pypdf>=5.1.0
beautifulsoup4>=4.12.3
requests>=2.32.4
python-docx>=1.1.2
python-pptx>=1.0.2
unstructured>=0.16.11

# Text Processing & NLP
transformers>=4.54.0
sentence-transformers>=5.0.0
tiktoken>=0.9.0
nltk>=3.9.1

# Vector Databases & Embeddings
qdrant-client>=1.13.2
faiss-cpu>=1.9.0


# Data Processing
pandas>=1.5.3
numpy>=1.26.4
python-dotenv>=1.0.0




